requirements.txt
fastapi
uvicorn
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
google-auth
requests
python-multipart
pydantic
